requirements.txt
undetected-chromedriver
seleniumbase
selenium-stealth